+++ core/src/main/kotlin/net/corda/core/flows/scheduler/mapper/FlowSchedulerMapper.kt
@@ -1,6 +1,7 @@
-package net.corda.core.flows
+package net.corda.core.flows.scheduler.mapper
 
 import net.corda.core.context.InvocationContext
+import net.corda.core.flows.FlowLogic
 import net.corda.core.identity.CordaX500Name
 
 /**
